Introduce GraphNode supertrait of TreeNode
